Add more explicit instructions for creation of yourusername.md in Intermediate and Advanced Route section
Step 4 in the Intermediate and Advanced Route is slightly confusing. The additional instructions
bring this section more in line with how this step is explained in the Beginner Route section (Step 4),
in case the user skipped over the Beginner Route and proceeded straight to the Intermediate and Advanced Route.

Some very minor typo edits in the Intermediate and Advanced Route section are also included.